Int ID移行: ビルドエラー修正完了
- GuildView: jobIndex→jobId、raceOrder辞書型修正
- StoryView: navigationDestination型をUInt16に変更
- ExplorationSnapshot: dungeonIdをUInt16に統一
- GemModificationProgressService: xxxIndex→xxxId命名統一
- InventoryProgressService: 型変換追加（UInt8/UInt16/Int）
- ItemPreloadService: xxxIndex→xxxId命名統一
- DropModels: SuperRareDailyState.jstDateをUInt32に変更
- ShopProgressService: 未使用変数警告修正
- 各種サービス・ビュー層の型整合性修正

ビルド成功（警告0件）

+++ Epika/Progress/Domain/Snapshots/ExplorationSnapshot.swift
@@ -54,7 +54,7 @@ struct ExplorationSnapshot: Sendable, Hashable {
 
     let persistentIdentifier: PersistentIdentifier
     var id: UUID
-    var dungeonId: UInt8
+    var dungeonId: UInt16
     var displayDungeonName: String
     var activeFloorNumber: Int
     var party: PartySummary
